requirements.txt
Flask>=2.0
mysql-connector-python>=8.0
psycopg2-binary>=2.9
gunicorn>=20.1.0
flask-cors>=3.0
python-dotenv>=0.19.0
